backend/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
twilio==8.10.3
openai==1.3.8
pydantic==2.5.0
python-multipart==0.0.6